<fix> : fix to warning C4244: 'argument' : conversion from 'std::streamsize' to 'unsigned int', possible loss of data

+++ Helper/ZlibHelper.cpp
@@ -45,7 +45,7 @@ BOOL CZlibHelper::Decompress(LPCTSTR inputFilePath, LPCTSTR outputFilePath)
 	inputFile.seekg(0, std::ios::beg);
 
 	// 입력 파일의 압축된 데이터를 메모리에 읽기
-	std::vector<char> compressedData(compressedSize);
+	std::vector<char> compressedData((int)compressedSize);
 	inputFile.read(compressedData.data(), compressedSize);
 
 	// zlib 스트림 준비
